add JWT validation options (#155)
Add environment variables EXPECTED_JWT_ISSUER and EXPECTED_JWT_AUDIENCE.
When these are both unset or empty, no validation of the JWT issuer and
audience claims is performed (this is the current behavior). When either
variable is non-empty, the corresponding JWT claim is validated to match
the value of that environment variable.